+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  * Change install & uninstall popups to reflect the new changes (d40f467)
  * Keep the scroll when enabling, installing etc. a mod (b3d444a)
  * Update MacOS crashlog to include base and offset (7816c43)
+ * Add user agent to AsyncWebRequest (c256207)
+ * Add post and custom requests to AsyncWebRequest (c256207)
 
 ## v1.2.1
  * Mods now target macOS 10.13 instead of 10.14 (7cc1cd4)